Add unit tests for Steam loading functionality
