+++ UnitTest/Tests/TestSteamLoader.cs
@@ -1,48 +1,20 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.Collections.Generic;
-using DoomLauncher.GameStores;
-using System.Linq;
-using static DoomLauncher.GameStores.StoreGameLoader;
+﻿using DoomLauncher.GameStores;
+using DoomLauncher.GameStores.Steam;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using System.IO;
 
 namespace UnitTest.Tests
 {
     [TestClass]
     public class TestSteamLoader
     {
-        
         [TestMethod]
-        public void LoadFromPath_FindsExpectedWads()
+        public void GetGameFolder()
         {
-            // Assuming fixtures:
+            var path = Path.GetFullPath(@"Resources\TestSteamInstall");
+            var gameFolder = SteamLoader.GetGameFolder(path, StoreGame.ULTIMATE_DOOM);
 
-            // SteamGame.ULTIMATE_DOOM:
-            // TestSteamLibrary1 / steamapps / common / TestDoom
-
-
-            GameFinder finder = _ => @"Resources\TestSteamLibrary1\steamapps\common\TestDoom";
-
-            var gameStoreFiles = StoreGameLoader.LoadStoreGame(StoreGame.ULTIMATE_DOOM, @"Resources\TestSteamLibrary1\steamapps\common\TestDoom");
-
-            var iwads = gameStoreFiles.InstalledIWads;
-            Assert.AreEqual(4, iwads.Count);
-            Assert.IsTrue(iwads.Exists(x => x.Contains("doom.wad")));
-            Assert.IsTrue(iwads.Exists(x => x.Contains("doom2.wad")));
-            Assert.IsTrue(iwads.Exists(x => x.Contains("plutonia.wad")));
-            Assert.IsTrue(iwads.Exists(x => x.Contains("tnt.wad")));
-
-            var pwads = gameStoreFiles.InstalledPWads;
-            Assert.AreEqual(4, pwads.Count);
-            Assert.IsTrue(pwads.Exists(x => x.Contains("nerve.wad")));
-            Assert.IsTrue(pwads.Exists(x => x.Contains("id1.wad")));
-            Assert.IsTrue(pwads.Exists(x => x.Contains("masterlevels.wad")));
-            Assert.IsTrue(pwads.Exists(x => x.Contains("sigil.wad")));
-        }
-
-        [TestMethod]
-        public void LoadFromPath_ReturnsEmptyFilesIfDirectoryNotFound()
-        {
-            var gameStoreFiles = StoreGameLoader.LoadStoreGame(StoreGame.ULTIMATE_DOOM, @"Resources\DoesNotExist");
-            Assert.AreEqual(gameStoreFiles, GameStoreFiles.EMPTY);
+            Assert.AreEqual(@"Resources\TestSteamLibrary1\steamapps\common\TestDoom", gameFolder);
         }
     }
 }
